NewsController refactor, deleted NewsDTO class and NewsDTOMapper.
