--- src/main/java/com/alkemy/ong/web/controller/mapper/NewsDTOMapper.java
@@ -1,32 +1,0 @@
-package com.alkemy.ong.web.controller.mapper;
-import com.alkemy.ong.domain.models.News;
-import com.alkemy.ong.web.controller.dto.NewsDTO;
-import org.springframework.stereotype.Component;
-
-@Component
-public class NewsDTOMapper {
-
-    public NewsDTO modelToDTO(News news){
-        NewsDTO returnDTO = NewsDTO.builder()
-                .id(news.getId())
-                .name(news.getName())
-                .content(news.getContent())
-                .image(news.getImage())
-                .categoryId(news.getCategoryId())
-                .type(news.getType())
-                .build();
-        return returnDTO;
-    }
-
-    public News DTOToModel(NewsDTO newsDTO){
-        News returnModel = News.builder()
-                .id(newsDTO.getId())
-                .name(newsDTO.getName())
-                .content(newsDTO.getContent())
-                .image(newsDTO.getImage())
-                .categoryId(newsDTO.getCategoryId())
-                .type(newsDTO.getType())
-                .build();
-        return returnModel;
-    }
-}
